25/07/2568 update file helloworld and variabels งานอาจารย์เบนย์ Part 2 ตอนบ่าย

+++ src/com/project/java_basic/ScannerBacis.java
@@ -4,9 +4,9 @@ import java.util.Scanner;
 
 public class ScannerBacis {
     public static void main(String[] args) {
-        Scanner MC= new Scanner(System.in);
+        Scanner MC = new Scanner(System.in);
         System.out.println("Input your number");
-        int number = MC.nextInt();
+        double number = MC.nextDouble();
         System.out.println("your number: "+ number);
     }
 }
